remove unnecessary boxing/unboxing
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1038303 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket/src/main/java/org/apache/wicket/protocol/http/ClientProperties.java
@@ -1128,7 +1128,7 @@ public class ClientProperties implements IClusterable
 
 				if (field.getType().equals(Integer.TYPE))
 				{
-					if (((Integer)value).intValue() == -1)
+					if ((Integer) value == -1)
 					{
 						value = null;
 					}
